LayoutTests/fast/url/ipv4-expected.txt

Test basic features of URL canonicalization

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canonicalize('http://192.168.0.1/') is 'http://192.168.0.1/'
FAIL canonicalize('http:///') should be http:///. Was http:/.
FAIL canonicalize('http://./') should be http:///. Was http://./.
PASS canonicalize('http://192.168.0.1/') is 'http://192.168.0.1/'
FAIL canonicalize('http://0300.0250.00.01/') should be http://192.168.0.1/. Was http://0300.0250.00.01/.
FAIL canonicalize('http://0xC0.0Xa8.0x0.0x1/') should be http://192.168.0.1/. Was http://0xC0.0Xa8.0x0.0x1/.
FAIL canonicalize('http://192.168.9.com/') should be http:///. Was http://192.168.9.com/.
FAIL canonicalize('http://19a.168.0.1/') should be http:///. Was http://19a.168.0.1/.
FAIL canonicalize('http://0308.0250.00.01/') should be http:///. Was http://0308.0250.00.01/.
FAIL canonicalize('http://0xCG.0xA8.0x0.0x1/') should be http:///. Was http://0xCG.0xA8.0x0.0x1/.
FAIL canonicalize('http://192/') should be http://0.0.0.192/. Was http://192/.
PASS successfullyParsed is true

TEST COMPLETE
